refactor web3 module
