+++ src/web3/siwe/siwe.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { SiweService } from './siwe.service';
+import { SiweController } from './siwe.controller';
+import { AccountModule } from '../../account/account.module';
+
+
+@Module({
+    imports: [AccountModule],
+    controllers: [SiweController],
+    providers: [SiweService],
+})
+export class SiweModule {}
