+++ src/web3/snapshot/snapshot.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { SnapshotService } from './snapshot.service';
+import { SnapshotController } from './snapshot.controller';
+
+@Module({
+    controllers: [SnapshotController],
+    providers: [SnapshotService],
+})
+export class SnapshotModule {}
