+++ src/web3/token-vote/ERC20.d.ts
